Clean Up Code + Fix Console Bug + Improve Keyword Count for "Complex-And" Searches where for example in one year, there were no cases but 4 of 5 of the keywords were used... get a count on those.
some function calls changed with this update...

From FSW_NAMELIST.py (RUN_ME folder) or whatever script you use to execute the Forecast Search Wizard, there is a new variable (default = False and does not need to be specified) which allows for bulk searches -- all warnings from the sub-iterations of the search are dumped to a single warning file rather than creating individual files for each element of the search. The output files are still individual. But now it is possible to pass the search criteria and output files directly to an analysis program OR return the files to a variable that can then be used locally depending on your setup.

As was the initial intention, minimal knowledge of python is required to use this tool. It is very much knobology. To perform some more complex searching and analysis, you may need to have to write or edit some code. Additional knobology was added for bulk searching and the FSW itself is now set up to handle these searches better. Another update with an updated script for RUN_ME and in tools will be coming soon.